Add Lasso and Ridge to linear regression methods
Lasso and Ridge added but it's clear that there is a bug in the
windowing of test and training windowed data: post-cv test set is
performing much worse than training data while this is not the case if I
split the training data into a new test and training set manually. To
investigate.